chore(resources): combine banner components
