+++ src/components/resource-sections/components/index.ts
@@ -5,8 +5,7 @@ import ResourceCitations from './citations';
 import RelatedDatasets from './related-datasets';
 import ResourceHeader from './header';
 import ResourceAuthors from './authors';
-import ResourceDates from './dates';
-import ResourceDatesAndOperatingSystems from './dates-and-operating-systems';
+import ResourceBanner from './banner';
 import ResourceOverview from './overview';
 import ResourceFilesTable from './files-table';
 import ResourceProvenance from './provenance';
@@ -17,8 +16,7 @@ export {
   ResourceAuthors,
   ResourceCitations,
   RelatedDatasets,
-  ResourceDates,
-  ResourceDatesAndOperatingSystems,
+  ResourceBanner,
   ResourceOverview,
   ResourceFilesTable,
   ResourceHeader,
